+++ src/packages/piral-ext/src/libs.ts
@@ -1,6 +1,4 @@
-export * from 'piral-fetch';
 export * from 'piral-translate';
-export * from 'piral-urql';
 export * from 'piral-dashboard';
 export * from 'piral-feeds';
 export * from 'piral-menu';